addLocation のロジックをカスタムフックに移動
